Paint Tool & Keybinds
Added an icon below the move tool to enable paint tool which currently paints all metal at once. Another way to access this is with the default keybind 3, an option to rekeybind it has been added.

There is currently not a save system for the paint so do not pull request as of this commit.

Added a keybind option for opening the sidebar (menu to add parts), default keybind is tab and this was done because some people could close the sidebar and then not reopen it until after protobot was restarted