Individually colorable metal
Redid the way that the ColorTool script was setting the color of the metal due to the save system being too good. (responsible for the changes in outlineselectionresponse script)

In the scenebuild script I added an if statement to check if a file was loaded from protobot 1.3.1,1.3.2, or 1.3.3 because new data was trying to be loaded that saves on those versions did not have.

As of right now this is ready for a pull request; however, you cannot color multiple pieces at the same time that is something i still need to work on

Also changed the version to 1.3.4 in build settings in order for the save system to work, can be changed to anything