Chain generator math done, kinda
I created a script that when a button is pressed it creates 3 game objects. 2 of which will be placed on either sprocket, the third gameobject holds the spline container which will be 0,0.
I added all the math and variables to find the coordinates where i need to place the spline knots, everything works except for the last two variables. i checked so many times. Its variables, "Upoint4" and "Ipoint4". Heres the desmos to anyone who wants to look into it, all the math is here (AND WORKING) https://www.desmos.com/calculator/6gon61wukl